Change handling of reply/qoute as it is possible that post has both
During implementation of qoute/reply posts I was sure that post cant be
qoute and a reply together in teh same time. It seems that it can so I changed
the behavior to prioritize replies to be able to download threads correctly.

Qoute posts are followed only when post is not replying to anything.