Always add caption with author and link to the media post
Tg allows you to have posts long for 4096 chars but media caption can only be
1024 chars. When post have a long description the description and media have
to be split in two messages. Before the change the media post had no caption,
now it gets a short caption with just an author and source link to make sure
author and source is always credited when eg. forwarding just a media post.