Covert booru webms to mp4
As for now there is no reencoding video and only the container is being changed
from webm to mp4. It is required because without it tg doesnt treat the file
as video and shows them as normal files. When the bot is deployed on more
powerful machine reencoding can be added as tg some devices running tg have
a hard time displaying files in vp9 and similar that are common in webm files.